feat(front): restructure callback components and add callback templates
